Move class DiscreteDerham to module psydac.api.feec
